refactor : 3주 1회차 패키지 구조 리팩토링
