+++ Junwon/test-code-with-architecture-main/src/main/java/com/example/demo/common/domain/exception/ResourceNotFoundException.java
@@ -1,0 +1,12 @@
+package com.example.demo.common.domain.exception;
+
+public class ResourceNotFoundException extends RuntimeException {
+
+    public ResourceNotFoundException(String datasource, long id) {
+        super(datasource + "에서 ID " + id + "를 찾을 수 없습니다.");
+    }
+
+    public ResourceNotFoundException(String datasource, String id) {
+        super(datasource + "에서 ID " + id + "를 찾을 수 없습니다.");
+    }
+}
