--- Junwon/test-code-with-architecture-main/src/main/java/com/example/demo/controller/ExceptionControllerAdvice.java
@@ -1,35 +1,0 @@
-package com.example.demo.controller;
-
-import static org.springframework.http.HttpStatus.FORBIDDEN;
-import static org.springframework.http.HttpStatus.NOT_FOUND;
-
-import com.example.demo.exception.CertificationCodeNotMatchedException;
-import com.example.demo.exception.ResourceNotFoundException;
-import lombok.RequiredArgsConstructor;
-import org.springframework.core.Ordered;
-import org.springframework.core.annotation.Order;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseBody;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ControllerAdvice
-@Order(Ordered.HIGHEST_PRECEDENCE)
-@RequiredArgsConstructor
-public class ExceptionControllerAdvice {
-
-    @ResponseBody
-    @ResponseStatus(NOT_FOUND)
-    @ExceptionHandler(ResourceNotFoundException.class)
-    public String resourceNotFoundException(ResourceNotFoundException exception) {
-        return exception.getMessage();
-    }
-
-    @ResponseBody
-    @ResponseStatus(FORBIDDEN)
-    @ExceptionHandler(CertificationCodeNotMatchedException.class)
-    public String certificationCodeNotMatchedException(CertificationCodeNotMatchedException exception) {
-        return exception.getMessage();
-    }
-
-}
