--- Junwon/test-code-with-architecture-main/src/main/java/com/example/demo/controller/HealthCheckController.java
@@ -1,18 +1,0 @@
-package com.example.demo.controller;
-
-import io.swagger.v3.oas.annotations.tags.Tag;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@Tag(name = "헬스 체크")
-@RestController
-public class HealthCheckController {
-
-    @GetMapping("/health_check.html")
-    public ResponseEntity<Void> healthCheck() {
-        return ResponseEntity
-            .ok()
-            .build();
-    }
-}
