--- Junwon/test-code-with-architecture-main/src/main/java/com/example/demo/controller/UserCreateController.java
@@ -1,33 +1,0 @@
-package com.example.demo.controller;
-
-import com.example.demo.model.dto.UserCreateDto;
-import com.example.demo.model.dto.UserResponse;
-import com.example.demo.repository.UserEntity;
-import com.example.demo.service.UserService;
-import io.swagger.v3.oas.annotations.tags.Tag;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@Tag(name = "유저(users)")
-@RestController
-@RequestMapping("/api/users")
-@RequiredArgsConstructor
-public class UserCreateController {
-
-    private final UserController userController;
-    private final UserService userService;
-
-    @PostMapping
-    public ResponseEntity<UserResponse> createUser(@RequestBody UserCreateDto userCreateDto) {
-        UserEntity userEntity = userService.create(userCreateDto);
-        return ResponseEntity
-            .status(HttpStatus.CREATED)
-            .body(userController.toResponse(userEntity));
-    }
-
-}
