--- Junwon/test-code-with-architecture-main/src/main/java/com/example/demo/exception/CertificationCodeNotMatchedException.java
@@ -1,8 +1,0 @@
-package com.example.demo.exception;
-
-public class CertificationCodeNotMatchedException extends RuntimeException {
-
-    public CertificationCodeNotMatchedException() {
-        super("자격 증명에 실패하였습니다.");
-    }
-}
