--- Junwon/test-code-with-architecture-main/src/main/java/com/example/demo/model/UserStatus.java
@@ -1,5 +1,0 @@
-package com.example.demo.model;
-
-public enum UserStatus {
-    PENDING, INACTIVE, ACTIVE
-}
