--- Junwon/test-code-with-architecture-main/src/main/java/com/example/demo/model/dto/MyProfileResponse.java
@@ -1,17 +1,0 @@
-package com.example.demo.model.dto;
-
-import com.example.demo.model.UserStatus;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class MyProfileResponse {
-
-    private Long id;
-    private String email;
-    private String nickname;
-    private String address;
-    private UserStatus status;
-    private Long lastLoginAt;
-}
