--- Junwon/test-code-with-architecture-main/src/main/java/com/example/demo/model/dto/PostResponse.java
@@ -1,15 +1,0 @@
-package com.example.demo.model.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class PostResponse {
-
-    private Long id;
-    private String content;
-    private Long createdAt;
-    private Long modifiedAt;
-    private UserResponse writer;
-}
