--- Junwon/test-code-with-architecture-main/src/main/java/com/example/demo/model/dto/PostUpdateDto.java
@@ -1,17 +1,0 @@
-package com.example.demo.model.dto;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-public class PostUpdateDto {
-
-    private final String content;
-
-    @Builder
-    public PostUpdateDto(
-        @JsonProperty("content") String content) {
-        this.content = content;
-    }
-}
