--- Junwon/test-code-with-architecture-main/src/main/java/com/example/demo/model/dto/UserResponse.java
@@ -1,16 +1,0 @@
-package com.example.demo.model.dto;
-
-import com.example.demo.model.UserStatus;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class UserResponse {
-
-    private Long id;
-    private String email;
-    private String nickname;
-    private UserStatus status;
-    private Long lastLoginAt;
-}
