--- Junwon/test-code-with-architecture-main/src/main/java/com/example/demo/model/dto/UserUpdateDto.java
@@ -1,20 +1,0 @@
-package com.example.demo.model.dto;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-public class UserUpdateDto {
-
-    private final String nickname;
-    private final String address;
-
-    @Builder
-    public UserUpdateDto(
-        @JsonProperty("nickname") String nickname,
-        @JsonProperty("address") String address) {
-        this.nickname = nickname;
-        this.address = address;
-    }
-}
