+++ Junwon/test-code-with-architecture-main/src/main/java/com/example/demo/post/controller/PostController.java
@@ -1,0 +1,51 @@
+package com.example.demo.post.controller;
+
+import com.example.demo.post.controller.response.PostResponse;
+import com.example.demo.post.domain.PostUpdate;
+import com.example.demo.post.infrastructure.PostEntity;
+import com.example.demo.post.service.PostService;
+import com.example.demo.user.controller.UserController;
+
+import io.swagger.v3.oas.annotations.tags.Tag;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@Tag(name = "게시물(posts)")
+@RestController
+@RequestMapping("/api/posts")
+@RequiredArgsConstructor
+public class PostController {
+
+    private final PostService postService;
+    private final UserController userController;
+
+    @GetMapping("/{id}")
+    public ResponseEntity<PostResponse> getPostById(@PathVariable long id) {
+        return ResponseEntity
+            .ok()
+            .body(toResponse(postService.getById(id)));
+    }
+
+    @PutMapping("/{id}")
+    public ResponseEntity<PostResponse> updatePost(@PathVariable long id, @RequestBody PostUpdate postUpdate) {
+        return ResponseEntity
+            .ok()
+            .body(toResponse(postService.update(id, postUpdate)));
+    }
+
+    public PostResponse toResponse(PostEntity postEntity) {
+        PostResponse PostResponse = new PostResponse();
+        PostResponse.setId(postEntity.getId());
+        PostResponse.setContent(postEntity.getContent());
+        PostResponse.setCreatedAt(postEntity.getCreatedAt());
+        PostResponse.setModifiedAt(postEntity.getModifiedAt());
+        PostResponse.setWriter(userController.toResponse(postEntity.getWriter()));
+        return PostResponse;
+    }
+}
