+++ Junwon/test-code-with-architecture-main/src/main/java/com/example/demo/post/controller/response/PostResponse.java
@@ -1,0 +1,17 @@
+package com.example.demo.post.controller.response;
+
+import com.example.demo.user.controller.response.UserResponse;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class PostResponse {
+
+    private Long id;
+    private String content;
+    private Long createdAt;
+    private Long modifiedAt;
+    private UserResponse writer;
+}
