+++ Junwon/test-code-with-architecture-main/src/main/java/com/example/demo/post/infrastructure/PostEntity.java
@@ -1,0 +1,39 @@
+package com.example.demo.post.infrastructure;
+
+import com.example.demo.user.infrastructure.UserEntity;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+@Entity
+@Table(name = "posts")
+public class PostEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(name = "content")
+    private String content;
+
+    @Column(name = "created_at")
+    private Long createdAt;
+
+    @Column(name = "modified_at")
+    private Long modifiedAt;
+
+    @ManyToOne
+    @JoinColumn(name = "user_id")
+    private UserEntity writer;
+
+}
