+++ Junwon/test-code-with-architecture-main/src/main/java/com/example/demo/post/infrastructure/PostRepository.java
@@ -1,0 +1,7 @@
+package com.example.demo.post.infrastructure;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PostRepository extends JpaRepository<PostEntity, Long> {
+
+}
