+++ Junwon/test-code-with-architecture-main/src/main/java/com/example/demo/post/service/PostService.java
@@ -1,0 +1,41 @@
+package com.example.demo.post.service;
+
+import com.example.demo.common.domain.exception.ResourceNotFoundException;
+import com.example.demo.post.domain.PostCreate;
+import com.example.demo.post.domain.PostUpdate;
+import com.example.demo.post.infrastructure.PostEntity;
+import com.example.demo.post.infrastructure.PostRepository;
+import com.example.demo.user.infrastructure.UserEntity;
+import com.example.demo.user.service.UserService;
+
+import java.time.Clock;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class PostService {
+
+    private final PostRepository postRepository;
+    private final UserService userService;
+
+    public PostEntity getById(long id) {
+        return postRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Posts", id));
+    }
+
+    public PostEntity create(PostCreate postCreate) {
+        UserEntity userEntity = userService.getById(postCreate.getWriterId());
+        PostEntity postEntity = new PostEntity();
+        postEntity.setWriter(userEntity);
+        postEntity.setContent(postCreate.getContent());
+        postEntity.setCreatedAt(Clock.systemUTC().millis());
+        return postRepository.save(postEntity);
+    }
+
+    public PostEntity update(long id, PostUpdate postUpdate) {
+        PostEntity postEntity = getById(id);
+        postEntity.setContent(postUpdate.getContent());
+        postEntity.setModifiedAt(Clock.systemUTC().millis());
+        return postRepository.save(postEntity);
+    }
+}
