--- Junwon/test-code-with-architecture-main/src/main/java/com/example/demo/repository/PostRepository.java
@@ -1,7 +1,0 @@
-package com.example.demo.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PostRepository extends JpaRepository<PostEntity, Long> {
-
-}
