--- Junwon/test-code-with-architecture-main/src/main/java/com/example/demo/repository/UserRepository.java
@@ -1,12 +1,0 @@
-package com.example.demo.repository;
-
-import com.example.demo.model.UserStatus;
-import java.util.Optional;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<UserEntity, Long> {
-
-    Optional<UserEntity> findByIdAndStatus(long id, UserStatus userStatus);
-
-    Optional<UserEntity> findByEmailAndStatus(String email, UserStatus userStatus);
-}
