+++ Junwon/test-code-with-architecture-main/src/main/java/com/example/demo/user/controller/UserCreateController.java
@@ -1,0 +1,33 @@
+package com.example.demo.user.controller;
+
+import com.example.demo.user.domain.UserCreate;
+import com.example.demo.user.controller.response.UserResponse;
+import com.example.demo.user.infrastructure.UserEntity;
+import com.example.demo.user.service.UserService;
+import io.swagger.v3.oas.annotations.tags.Tag;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@Tag(name = "유저(users)")
+@RestController
+@RequestMapping("/api/users")
+@RequiredArgsConstructor
+public class UserCreateController {
+
+    private final UserController userController;
+    private final UserService userService;
+
+    @PostMapping
+    public ResponseEntity<UserResponse> createUser(@RequestBody UserCreate userCreate) {
+        UserEntity userEntity = userService.create(userCreate);
+        return ResponseEntity
+            .status(HttpStatus.CREATED)
+            .body(userController.toResponse(userEntity));
+    }
+
+}
