+++ Junwon/test-code-with-architecture-main/src/main/java/com/example/demo/user/controller/response/UserResponse.java
@@ -1,0 +1,16 @@
+package com.example.demo.user.controller.response;
+
+import com.example.demo.user.domain.UserStatus;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class UserResponse {
+
+    private Long id;
+    private String email;
+    private String nickname;
+    private UserStatus status;
+    private Long lastLoginAt;
+}
