+++ Junwon/test-code-with-architecture-main/src/main/java/com/example/demo/user/domain/MyProfileResponse.java
@@ -1,0 +1,17 @@
+package com.example.demo.user.domain;
+
+import com.example.demo.user.domain.UserStatus;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class MyProfileResponse {
+
+    private Long id;
+    private String email;
+    private String nickname;
+    private String address;
+    private UserStatus status;
+    private Long lastLoginAt;
+}
