+++ Junwon/test-code-with-architecture-main/src/main/java/com/example/demo/user/domain/UserStatus.java
@@ -1,0 +1,5 @@
+package com.example.demo.user.domain;
+
+public enum UserStatus {
+    PENDING, INACTIVE, ACTIVE
+}
