+++ Junwon/test-code-with-architecture-main/src/main/java/com/example/demo/user/domain/UserUpdate.java
@@ -1,0 +1,20 @@
+package com.example.demo.user.domain;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+public class UserUpdate {
+
+    private final String nickname;
+    private final String address;
+
+    @Builder
+    public UserUpdate(
+        @JsonProperty("nickname") String nickname,
+        @JsonProperty("address") String address) {
+        this.nickname = nickname;
+        this.address = address;
+    }
+}
