+++ Junwon/test-code-with-architecture-main/src/main/java/com/example/demo/user/infrastructure/UserRepository.java
@@ -1,0 +1,12 @@
+package com.example.demo.user.infrastructure;
+
+import com.example.demo.user.domain.UserStatus;
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<UserEntity, Long> {
+
+    Optional<UserEntity> findByIdAndStatus(long id, UserStatus userStatus);
+
+    Optional<UserEntity> findByEmailAndStatus(String email, UserStatus userStatus);
+}
