+++ Junwon/test-code-with-architecture-main/src/test/java/com/example/demo/controller/PostCreateControllerTest.java
@@ -14,7 +14,7 @@ import org.springframework.test.context.jdbc.Sql.ExecutionPhase;
 import org.springframework.test.context.jdbc.SqlGroup;
 import org.springframework.test.web.servlet.MockMvc;
 
-import com.example.demo.model.dto.PostCreateDto;
+import com.example.demo.post.domain.PostCreate;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 @SpringBootTest
@@ -33,7 +33,7 @@ public class PostCreateControllerTest {
 	@Test
 	void 사용자는_게시물을_작성할_수_있다() throws Exception {
 		// given
-		PostCreateDto postCreateDto = PostCreateDto.builder()
+		PostCreate postCreate = PostCreate.builder()
 			.writerId(1)
 			.content("helloworld")
 			.build();
@@ -43,7 +43,7 @@ public class PostCreateControllerTest {
 		mockMvc.perform(
 				post("/api/posts")
 					.contentType(MediaType.APPLICATION_JSON)
-					.content(objectMapper.writeValueAsString(postCreateDto)))
+					.content(objectMapper.writeValueAsString(postCreate)))
 			.andExpect(status().isCreated())
 			.andExpect(jsonPath("$.id").isNumber())
 			.andExpect(jsonPath("$.content").value("helloworld"))
