+++ Junwon/test-code-with-architecture-main/src/test/java/com/example/demo/controller/UserCreateControllerTest.java
@@ -18,7 +18,7 @@ import org.springframework.test.context.jdbc.SqlGroup;
 import org.springframework.test.web.servlet.MockMvc;
 import static org.mockito.ArgumentMatchers.any;
 
-import com.example.demo.model.dto.UserCreateDto;
+import com.example.demo.user.domain.UserCreate;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 @SpringBootTest
@@ -40,7 +40,7 @@ public class UserCreateControllerTest {
 	@Test
 	void 사용자는_회원_가입을_할_수_있고_회원가입된_사용자는_PENDING_상태이다() throws Exception {
 		// given
-		UserCreateDto userCreateDto = UserCreateDto.builder()
+		UserCreate userCreate = UserCreate.builder()
 			.email("kok202@naver.com")
 			.nickname("kok202")
 			.address("Pangyo").build();
@@ -51,7 +51,7 @@ public class UserCreateControllerTest {
 		mockMvc.perform(post("/api/users")
 				.header("EMAIL", "kok202@naver.com")
 				.contentType(MediaType.APPLICATION_JSON)
-				.content(objectMapper.writeValueAsString(userCreateDto)))
+				.content(objectMapper.writeValueAsString(userCreate)))
 			.andExpect(status().isCreated())
 			.andExpect(jsonPath("$.id").value(1))
 			.andExpect(jsonPath("$.email").value("kok202@naver.com"))
