+++ Junwon/test-code-with-architecture-main/src/test/java/com/example/demo/repository/UserRepositoryTest.java
@@ -10,7 +10,9 @@ import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.jdbc.Sql;
 
-import com.example.demo.model.UserStatus;
+import com.example.demo.user.domain.UserStatus;
+import com.example.demo.user.infrastructure.UserEntity;
+import com.example.demo.user.infrastructure.UserRepository;
 
 @DataJpaTest(showSql = true)
 @Sql("/sql/user-repository-test-data.sql")
