+++ Junwon/test-code-with-architecture-main/src/test/java/com/example/demo/service/PostServiceTest.java
@@ -12,9 +12,10 @@ import org.springframework.test.context.jdbc.Sql;
 import org.springframework.test.context.jdbc.Sql.ExecutionPhase;
 import org.springframework.test.context.jdbc.SqlGroup;
 
-import com.example.demo.model.dto.PostCreateDto;
-import com.example.demo.model.dto.PostUpdateDto;
-import com.example.demo.repository.PostEntity;
+import com.example.demo.post.domain.PostCreate;
+import com.example.demo.post.domain.PostUpdate;
+import com.example.demo.post.infrastructure.PostEntity;
+import com.example.demo.post.service.PostService;
 
 @SpringBootTest
 @TestPropertySource("classpath:test-application.properties")
@@ -44,13 +45,13 @@ public class PostServiceTest {
 	@Test
 	void postCreateDto_를_이용하여_게시물을_생성할_수_있다(){
 		// given
-		PostCreateDto postCreateDto = PostCreateDto.builder()
+		PostCreate postCreate = PostCreate.builder()
 				.writerId(1)
 				.content("foobar")
 				.build();
 
 		// when
-		PostEntity result = postService.create(postCreateDto);
+		PostEntity result = postService.create(postCreate);
 
 		// then
 		assertThat(result.getId()).isNotNull();
@@ -61,12 +62,12 @@ public class PostServiceTest {
 	@Test
 	void postUpdateDto_를_이용하여_게시물을_수정할_수_있다(){
 		// given
-		PostUpdateDto postUpdateDto = PostUpdateDto.builder()
+		PostUpdate postUpdate = PostUpdate.builder()
 			.content("hello world :)")
 			.build();
 
 		// when
-		postService.update(1, postUpdateDto);
+		postService.update(1, postUpdate);
 
 		// then
 		PostEntity postEntity = postService.getById(1);
