+++ Junwon/test-code-with-architecture-main/src/test/java/com/example/demo/service/UserServiceTest.java
@@ -15,12 +15,13 @@ import org.springframework.test.context.jdbc.Sql;
 import org.springframework.test.context.jdbc.Sql.ExecutionPhase;
 import org.springframework.test.context.jdbc.SqlGroup;
 
-import com.example.demo.exception.CertificationCodeNotMatchedException;
-import com.example.demo.exception.ResourceNotFoundException;
-import com.example.demo.model.UserStatus;
-import com.example.demo.model.dto.UserCreateDto;
-import com.example.demo.model.dto.UserUpdateDto;
-import com.example.demo.repository.UserEntity;
+import com.example.demo.common.domain.exception.CertificationCodeNotMatchedException;
+import com.example.demo.common.domain.exception.ResourceNotFoundException;
+import com.example.demo.user.domain.UserStatus;
+import com.example.demo.user.domain.UserCreate;
+import com.example.demo.user.domain.UserUpdate;
+import com.example.demo.user.infrastructure.UserEntity;
+import com.example.demo.user.service.UserService;
 
 @SpringBootTest
 @TestPropertySource("classpath:test-application.properties")
@@ -83,7 +84,7 @@ public class UserServiceTest {
 	@Test
 	void userCreateDto_를_이용하여_유저를_생성할_수_있다(){
 		// given
-		UserCreateDto userCreateDto = UserCreateDto.builder()
+		UserCreate userCreate = UserCreate.builder()
 			.email("kok202@kakao.com")
 			.address("Gyeongi")
 			.nickname("kok202-k")
@@ -92,7 +93,7 @@ public class UserServiceTest {
 		BDDMockito.doNothing().when(javaMailSender).send(any(SimpleMailMessage.class));
 
 		// when
-		UserEntity result = userService.create(userCreateDto);
+		UserEntity result = userService.create(userCreate);
 
 		// then
 		assertThat(result.getId()).isNotNull();
@@ -103,13 +104,13 @@ public class UserServiceTest {
 	@Test
 	void userUpdateDto_를_이용하여_유저를_수정할_수_있다(){
 		// given
-		UserUpdateDto userUpdateDto = UserUpdateDto.builder()
+		UserUpdate userUpdate = UserUpdate.builder()
 			.address("Incheon")
 			.nickname("kok202-n")
 			.build();
 
 		// when
-		userService.update(1, userUpdateDto);
+		userService.update(1, userUpdate);
 
 		// then
 		UserEntity userEntity = userService.getById(1);
